requirements.txt
requests>=2.22.0
jsonschema==4.19.2
Click>=7.0
tabulate==0.9.0
pyyaml==6.0.1
junit-xml==1.9
